Improve shard splitting.
Previously, the __scatter__ query would be attempted using all the
filters specified for the mapreduce. If there was no index or no keys
returned, then it would immediately fall back to querying with no
filters, leading to inefficient sharding. This case was fairly common
and impossible to predict because it was based on whether the filters
would filter out all entities with the __scatter__ property.

Instead, now, the full query will be attempted and if there is no index
or no keys returned, then a filter will be removed and the query will
be retried. This will continue until either at least on key is returned
or we remove all the filters. At that point we will fall back to
querying with no filters as before.

Note that this means that the order that the filters are added is
significant since at each step the last filter will be removed. This can
be used to tune how the query space expands at each step.